FastAPI重构/fastapi-main/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库相关
sqlalchemy==1.4.37
pymysql==1.0.2
alembic==1.12.1

# 认证相关
authlib==1.0.1
werkzeug==2.1.2

# 定时任务
apscheduler==3.9.1

# HTTP 请求
requests==2.27.1

# 数据处理
pandas==1.3.5
numpy==1.21.6

# 配置管理
pyyaml==6.0

# 爬虫相关
beautifulsoup4==4.10.0
lxml==4.9.1

# 工具库
typing-extensions==4.8.0
python-multipart==0.0.6

# 开发工具（可选）
# pytest==7.4.0
# pytest-asyncio==0.21.1
# black==22.6.0
# flake8==5.0.4
